Conditionally remove file

+++ scripts/bg.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
-rm /tmp/$1.log
+if [ -f "/tmp/$1.log" ]; then
+    rm /tmp/$1.log
+fi 
 $@ > /tmp/$1.log 2>&1 &
